pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "firecrawl-py>=2.8.0",
    "langchain-aws>=0.2.25",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
]
